Refactor: change instance 'user' to 'session'
