+++ routes/index.js
@@ -1,11 +1,11 @@
 import express from 'express';
 
-import userRouter from './userRouter.js';
+import sessionRouter from './sessionRouter.js';
 import authRouter from './authRouter.js';
 
 const router = express.Router();
 
-router.use(userRouter);
+router.use(sessionRouter);
 router.use(authRouter);
 
 export default router;
